requirements.txt
Django==1.9
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.0.0
PyYAML==3.11
Pygments==2.0.2
Twisted==15.4.0
Werkzeug==0.10.4
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==2015.9.6.2
cffi==0.8.6
chardet==2.3.0
click==5.1
colorama==0.3.2
configobj==5.0.6
cryptography==0.6.1
django-fluent-pages==0.9
django-fluent-utils==1.2.1
django-mptt==0.7.4
django-parler==1.5.1
django-polymorphic==0.7.2
django-polymorphic-tree==1.1.1
django-tag-parser==2.1
django-wysiwyg==0.7.1
django-wysiwyg-redactor==0.4.9
future==0.15.2
html5lib==0.999
itsdangerous==0.24
lxml==3.4.0
mercurial==3.1.2
ndg-httpsclient==0.3.2
dj-database-url==0.3.0
gunicorn==19.1.1
dj-static==0.0.6
psycopg2==2.5.1
